fix: price format

+++ app/src/main/java/com/erickpimentel/mercadoeditorial/adapter/BookPagingDataAdapter.kt
@@ -25,10 +25,11 @@ class BookPagingDataAdapter @Inject constructor(): PagingDataAdapter<Book, BookP
 
     inner class BookViewHolder(private val binding: BookViewBinding): RecyclerView.ViewHolder(binding.root) {
         fun bind(holder: BookViewHolder, book: Book){
+
             binding.apply {
                 bookTitle.text = book.titulo
                 bookType.text = book.formato
-                bookPrice.text = holder.itemView.context.resources.getString(R.string.price_symbol, book.preco)
+                bookPrice.text = holder.itemView.context.resources.getString(R.string.price_symbol, book.preco.toFloat())
                 bookImageView.load(book.imagens.imagem_primeira_capa.pequena){
                     crossfade(true)
                     placeholder(R.drawable.placeholder)
